+++ src/solver/fold/fold_backward.cpp
@@ -61,15 +61,15 @@ ConvSolution FoldBwd::GetSolution([[maybe_unused]] const ExecutionContext& conte
     auto input_grad_dims  = problem.GetDinputDesc().GetLengths();
     auto output_grad_dims = problem.GetDoutputDesc().GetLengths();
 
-    const int32_t N          = static_cast<int32_t>(output_grad_dims[0]);
-    const int32_t C          = static_cast<int32_t>(output_grad_dims[1]);
-    int32_t spatial_dim_size = output_grad_dims.size() - 2;
-    int32_t P = 1, L = 1;
-    std::vector<int32_t> ls;
+    const int64_t N          = static_cast<int64_t>(output_grad_dims[0]);
+    const int64_t C          = static_cast<int64_t>(output_grad_dims[1]);
+    int64_t spatial_dim_size = output_grad_dims.size() - 2;
+    int64_t P = 1, L = 1;
+    std::vector<int64_t> ls;
     for(int i = 0; i < spatial_dim_size; ++i)
     {
         P *= problem.kernel_size[i];
-        int32_t l = (static_cast<int32_t>(output_grad_dims[i + 2]) + 2 * problem.padding[i] -
+        int64_t l = (static_cast<int64_t>(output_grad_dims[i + 2]) + 2 * problem.padding[i] -
                      problem.dilation[i] * (problem.kernel_size[i] - 1) - 1) /
                         problem.stride[i] +
                     1;
@@ -90,7 +90,7 @@ ConvSolution FoldBwd::GetSolution([[maybe_unused]] const ExecutionContext& conte
     kernel.comp_options = build_params.GenerateFor(kbp::HIP{});
 
     size_t xlocalsize = LOCAL_SIZE;
-    size_t xgridsize  = AlignUp(N * C * P * L, LOCAL_SIZE);
+    size_t xgridsize  = AlignUp(static_cast<size_t>(N * C * P * L), LOCAL_SIZE);
     size_t ylocalsize = 1;
     size_t ygridsize  = 1;
     size_t zlocalsize = 1;
@@ -115,18 +115,18 @@ ConvSolution FoldBwd::GetSolution([[maybe_unused]] const ExecutionContext& conte
             auto input_grad_dims  = deref(params.dinputDesc).GetLengths();
             auto output_grad_dims = deref(params.doutputDesc).GetLengths();
 
-            int32_t kernel_size_h = params.kernel_size[0];
-            int32_t kernel_size_w = params.kernel_size[1];
-            int32_t stride_h      = params.stride[0];
-            int32_t stride_w      = params.stride[1];
-            int32_t padding_h     = params.padding[0];
-            int32_t padding_w     = params.padding[1];
-            int32_t dilation_h    = params.dilation[0];
-            int32_t dilation_w    = params.dilation[1];
-            int32_t LH            = ls[0];
-            int32_t LW            = ls[1];
-            int32_t H             = static_cast<int32_t>(output_grad_dims[2]);
-            int32_t W             = static_cast<int32_t>(output_grad_dims[3]);
+            int64_t kernel_size_h = params.kernel_size[0];
+            int64_t kernel_size_w = params.kernel_size[1];
+            int64_t stride_h      = params.stride[0];
+            int64_t stride_w      = params.stride[1];
+            int64_t padding_h     = params.padding[0];
+            int64_t padding_w     = params.padding[1];
+            int64_t dilation_h    = params.dilation[0];
+            int64_t dilation_w    = params.dilation[1];
+            int64_t LH            = ls[0];
+            int64_t LW            = ls[1];
+            int64_t H             = static_cast<int64_t>(output_grad_dims[2]);
+            int64_t W             = static_cast<int64_t>(output_grad_dims[3]);
 
             kernel(params.doutput,
                    params.dinput,
