+++ src/solver/fold/fold_forward.cpp
@@ -61,10 +61,10 @@ ConvSolution FoldFwd::GetSolution([[maybe_unused]] const ExecutionContext& conte
     auto input_dims = problem.GetInputDesc().GetLengths();
 
     auto output_dims = problem.GetOutputDesc().GetLengths();
-    const int32_t N  = static_cast<int32_t>(output_dims[0]);
-    const int32_t C  = static_cast<int32_t>(output_dims[1]);
-    int32_t H        = static_cast<int32_t>(output_dims[2]);
-    int32_t W        = static_cast<int32_t>(output_dims[3]);
+    const int64_t N  = static_cast<int64_t>(output_dims[0]);
+    const int64_t C  = static_cast<int64_t>(output_dims[1]);
+    int64_t H        = static_cast<int64_t>(output_dims[2]);
+    int64_t W        = static_cast<int64_t>(output_dims[3]);
 
     {
         auto kernel        = KernelInfo{};
@@ -80,7 +80,7 @@ ConvSolution FoldFwd::GetSolution([[maybe_unused]] const ExecutionContext& conte
         kernel.comp_options = build_params.GenerateFor(kbp::HIP{});
 
         size_t xlocalsize = LOCAL_SIZE;
-        size_t xgridsize  = AlignUp(N * C * H * W, LOCAL_SIZE);
+        size_t xgridsize  = AlignUp(static_cast<size_t>(N * C * H * W), LOCAL_SIZE);
         size_t ylocalsize = 1;
         size_t ygridsize  = 1;
         size_t zlocalsize = 1;
@@ -106,13 +106,13 @@ ConvSolution FoldFwd::GetSolution([[maybe_unused]] const ExecutionContext& conte
             auto input_dims  = deref(params.inputDesc).GetLengths();
             auto output_dims = deref(params.outputDesc).GetLengths();
 
-            int32_t spatial_dim_size = output_dims.size() - 2;
-            int32_t P = 1, L = 1;
-            std::vector<int32_t> ls;
+            int64_t spatial_dim_size = output_dims.size() - 2;
+            int64_t P = 1, L = 1;
+            std::vector<int64_t> ls;
             for(int i = 0; i < spatial_dim_size; ++i)
             {
                 P *= params.kernel_size[i];
-                int32_t l = (static_cast<int32_t>(output_dims[i + 2]) + 2 * params.padding[i] -
+                int64_t l = (static_cast<int64_t>(output_dims[i + 2]) + 2 * params.padding[i] -
                              params.dilation[i] * (params.kernel_size[i] - 1) - 1) /
                                 params.stride[i] +
                             1;
@@ -120,18 +120,18 @@ ConvSolution FoldFwd::GetSolution([[maybe_unused]] const ExecutionContext& conte
                 ls.push_back(l);
             }
 
-            int32_t kernel_size_h = params.kernel_size[0];
-            int32_t kernel_size_w = params.kernel_size[1];
-            int32_t stride_h      = params.stride[0];
-            int32_t stride_w      = params.stride[1];
-            int32_t padding_h     = params.padding[0];
-            int32_t padding_w     = params.padding[1];
-            int32_t dilation_h    = params.dilation[0];
-            int32_t dilation_w    = params.dilation[1];
-            int32_t LH            = ls[0];
-            int32_t LW            = ls[1];
-            int32_t H             = static_cast<int32_t>(output_dims[2]);
-            int32_t W             = static_cast<int32_t>(output_dims[3]);
+            int64_t kernel_size_h = params.kernel_size[0];
+            int64_t kernel_size_w = params.kernel_size[1];
+            int64_t stride_h      = params.stride[0];
+            int64_t stride_w      = params.stride[1];
+            int64_t padding_h     = params.padding[0];
+            int64_t padding_w     = params.padding[1];
+            int64_t dilation_h    = params.dilation[0];
+            int64_t dilation_w    = params.dilation[1];
+            int64_t LH            = ls[0];
+            int64_t LW            = ls[1];
+            int64_t H             = static_cast<int64_t>(output_dims[2]);
+            int64_t W             = static_cast<int64_t>(output_dims[3]);
 
             kernel(params.input,
                    params.output,
